fix fabric bug with reporting bw telemetry in 2-erisc mode (#31602)
In 2-erisc mode in fabric, when we are accumulating words/packets sent
for bandwidth telemetry, we no longer accumulate for the receiver side,
because it does not send payloads. This was causing an issue of cycles
being double counted (once for erisc0 and once for erisc1), but BW only
being counted once.

### Checklist
- [x] T3K Quick:
https://github.com/tenstorrent/tt-metal/actions/runs/18996937581
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable):
https://github.com/tenstorrent/tt-metal/actions/runs/18998026994
- [x] BH multi device unit test:
https://github.com/tenstorrent/tt-metal/actions/runs/18983015276